Removed check which was refusing to check for extensions in SSL 3.0
Sniproxy contained a check that would refuse to parse an SSL 3.0 packet
for extensions. This check was wrong as SSL 3.0 explicitly allows for
additional data in the end of a Client Hello (see RFC6101 last paragraph of
5.6.1.2). In fact most browsers set SSL 3.0 as the record packet version
irrespectively of the TLS version they support (which is advertized inside
the Client Hello).

Signed-off-by: Nikos Mavrogiannopoulos <nmav@redhat.com>