Fallback on cross address family transparent proxy
When configured for transparent mode it is not possible use an IPv6
source address when connecting to an IPv4 backend server or vice versa.
If the address families differ, fall back to using the proxy hosts
address for the connection to the backend server. When resolving backend
server hostnames in transparent mode, only consider the resource records
of the matching address family.