DTLS Protocol Support
This adds support for processing DTLS packets, specifically to acquire the
SNI value from the ClientHello packet. DTLS and TLS packets are different
enough that a separate parser was required. However, it appears extension
processing is the same. So I've moved both parse_extensions() and
parse_server_name_extension() into new sni.[c,h] files so those functions
can be shared by both the tls and dtls processors.

Note that this code is untested at the moment.

Related to issue #352.

Signed-off-by: Kyle Mestery <mestery@mestery.com>