Update buffer module to work with UDP datagrams
The buffer code was meant to work with stream sockets. This commits modifies it
so it works with datagram sockets as well as stream sockets.

Proxying datagram sockets requires the ability to push an exact copy of each
frame from source to destination and vice versa. Thus, we introduce the
ability for the Buffer structure to understand if it is associated with a
SOCK_STREAM or a SOCK_DGRAM. If a SOCK_DGRAM is being used, for each write
into the buffer we first reserver two bytes and we store the length. When
reading, we first read the length as well.

Tests were updated to test the new SOCK_DGRAM functionality.

Note that when testing this locally it doesn't quite work yet. I am pushing
this commit out so it can be reviewed while I debug it.

Signed-off-by: Kyle Mestery <mestery@mestery.com>