buffer: Make buffer_coalesce more intelligent
For the datagram case, we need to replicate the exact buffer we're
coalescing, including the individual datagram sizes. The code as it
existed before this patch was writing the entire buffer length into
the first 2 bytes of datagram size, messing it all up. This fixes
it to iterate the existing Buffer, copying by pieces into a new
buffer, and then copying the resulting buffer back over.

Signed-off-by: Kyle Mestery <mestery@mestery.com>